fix: handle multiple include/exclude paths with newlines
When multiple paths are specified using the `|` operator in YAML, the resulting string contains newlines.
When this string is passed to `globs=$(eval "echo ${file_paths}")` bash is trying to execute paths as commend and fails with the error:

```sh
/action/functions.sh: line 215: tests/**: No such file or directory
```

We can avoid this error by replacing newlines with spaces from the paths.