feat: improve console output by using `csgrep -U`
Show more context for ShellCheck defects and fixes in console output.
Using the `csgrep --embed-context` option allows us to show a part of
the code in which a defect is reported. This is much more user-friendly
than the current plain GCC output format.

This commit also fixes issues with statistics calculations.

Related to: https://github.com/redhat-plumbers-in-action/differential-shellcheck/issues/88